Add files via upload
Title: Hospital Management System

Description: This project is a simple CRUD (Create, Read, Update, Delete) application built in Java with MySQL as the database, designed to manage data for a hospital. The system allows authorized users to perform basic data management operations, such as adding, updating, and deleting patient and doctor records, and searching for specific records by various parameters. The interface is designed to be user-friendly and intuitive, with easy navigation between different sections of the application. The database is structured to ensure data consistency and integrity, and the application is built using industry-standard coding practices and libraries. This project can be useful for hospitals and healthcare facilities looking for a simple and customizable data management solution.